cambie el tipo a double, ya que las funciones para calcular las distancias son double

+++ src/main/java/com/ddsutn/group01/tpanual/models/pois/Coordenada.java
@@ -2,19 +2,19 @@ package com.ddsutn.group01.tpanual.models.pois;
 
 public class Coordenada {
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Coordenada (int valor1, int valor2) {
+	public Coordenada (double valor1, double valor2) {
 		this.x = valor1;
 		this.y = valor2;
 	}
 	
-	public int getX () {
+	public double getX () {
 		return x;
 	}
 	
-	public int getY () {
+	public double getY () {
 		return y;
 	}
 }
